step1_video_highlight_extractor/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
psycopg2-binary==2.9.9
pgvector==0.2.3
python-dotenv==1.0.0
google-generativeai==0.3.2
SpeechRecognition==3.10.0
moviepy==1.0.3
python-magic==0.4.27
SQLAlchemy==2.0.23
sqlalchemy-utils==0.41.1
pydantic==2.5.2
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
scenedetect==0.6.2
faster-whisper==0.10.0
torch==2.1.2
torchvision==0.16.2
ultralytics==8.0.196  # For YOLOv8
librosa==0.10.1
soundfile==0.12.1
langchain-google-genai==0.0.5
langchain-core==0.1.17 